fix(schemas): scope BillingStatus types to @useatlas/schemas + address PR #1678 review
Scaffold smoke-tests build against @useatlas/types@0.0.11 from npm — which
doesn't export OVERAGE_STATUSES or the BillingStatus family yet. Moving
those newly-added symbols from @useatlas/types into @useatlas/schemas
keeps the workspace build green without needing a types package bump +
two-PR publish dance.

Changes:
- Revert @useatlas/types billing additions (OVERAGE_STATUSES tuple,
  BillingStatus + 6 nested interfaces). OverageStatus union stays.
- Move BillingStatus + nested interfaces to
  packages/schemas/src/billing.ts (the module that now owns the wire
  contract anyway). `satisfies z.ZodType<OverageStatus>` on the local
  OVERAGE_STATUSES tuple keeps compile-time drift detection against
  the types-package union.
- packages/web/src/app/admin/billing/page.tsx imports BillingStatus
  from @useatlas/schemas instead.
- Add ATLAS_ROLES canonical-tuple assertion + per-role parse coverage
  (pr-test-analyzer nit, closes the gap surfaced during review).

Local billing tuple has no behavior change — the tests still pin the
exact values and the satisfies clause still catches drift.
